test(di): avoid pytest loader when importing modules
We avoid the use of the pytest loader when importing a module in one of
the DI tests for decorated function discovery. Because the current
implementation of the pytest loader does not implement the get_code
method, we cannot wrap around it to get the code object for the module
before it is executed. As a workaround in tests, we temporarily remove
the pytest loader so that we can import the module under "normal"
conditions. Once the test is done we restore the state of the
sys.meta_path list.