chore(debugging): guard against race on emitting status
We guard the probe registry against potential race conditions when
setting the `EMITTING` status. This can happen if a probe that has been
instrumented is being deleted, but the instrumentation is triggered in a
user thread at around the same time the RC client relies the
configuration payload. In this case we simply log the event and do
nothing.

The better solution would be to lock the registry access, since the
handling of the `EMITTING` status exposes the registry to concurrent
access. However, we opted for not introducing a lock to avoid adding
blocking in-line on user code.