fix(ci_visibility): count failed/skipped tests in JUnit XML when retries are enabled (#12862)
The pytest JUnit XML plugin uses the test report's
[`failed`](https://github.com/pytest-dev/pytest/blob/8.3.x/src/_pytest/junitxml.py#L562)
and
[`longrepr`](https://github.com/pytest-dev/pytest/blob/8.3.x/src/_pytest/junitxml.py#L201)
properties to count failed tests and include them in the output. Because
retried tests have their own special statuses (`dd_efd_final_failed`,
etc), they don't count as failures, and are excluded from the JUnit XML
count. This PR creates a subclass of TestReport that is aware of those
special statuses and reports them as passed/failed/skipped accordingly.

This is honestly a bit of a hack. It would probably be best to rewrite
the retry logic entirely so it would use normal pytest states, and pass
the information that they are retries in some other way. But that will
take more time, and I would like to fix the bug sooner rather than
later.

The exception information for the initial attempt is included in the JUnit XML.

Known issue: quarantined failing tests are not counted. The way forward
with this is to rewrite the retry logic, which I plan to do in a future
PR.

- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
(cherry picked from commit 2a506fb25a1bb5791f703329624cd18858dd357d)