ci: add pip cache to CI jobs (#12876)
Adds a pip cache to a large number of our CI jobs.

For riot: this does this by using a unique pip cache key per-job based
on the sha256sum of all of the `.riot/requirements/*.txt` files for all
the venvs matching the suite name or pattern. This way if any deps
change we'll invalidate the cache to ensure it doesn't grow in size
forever.

For hatch: this uses the job name as the key.

The goal is to cache the downloaded wheels/distributions to save on
network bandwidth and effort for downloading every time. This will also
help cache built wheels if the downloaded distribution is a source
distribution.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)