refactor: configuration instance
We move the creation of the configuration instance in the module where
the class itself is defined. We then update the imports to grab a
reference directly from the settings module, instead of the root module
ddtrace. This way we remove a substantial number of circular import
paths within the library.