fix(openai): [MLOB-2454] avoid creating spans for with_raw_response streamed requests (#12874)
This PR addresses a bug with the Open AI integration for streamed
requests using the `with_raw_response` wrapper. The issue here is that
the stream parsed from the `with_raw_response` response will not be
traced by our integration. Therefore, even if the user iterates over
this stream, our traced stream will never be consumed and the span
created from this request will never be finished, leading to errors and
dropped spans.

Since we cannot trace the stream parsed after the raw response is
returned to the user, the solution here is to avoid creating any spans
that will not be finished (e.g. spans for streaming requests that use
`with_raw_response`). This is done by modifying the kwargs to include a
`_dd.with_raw_response` flag which is set to True when the raw response
is requested for a streamed request.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)