feat: add service discovery (#12809)
# Description

This PR implements a new mechanism for storing tracer configuration
using memfd. By leveraging this approach, we establish a reliable and
secure way to monitor which processes are instrumented, enhancing
Service Discovery and improving visibility into tracing library usage.

### Why this change?
The current method of inferring instrumented processes by analyzing
command-line arguments is fragile, inconsistent and prone to false
positives/negatives. With this update, each `Tracer` instance will
publish a subset of the configuration that can be used as a source of
truth, allowing the Datadog Agent to reliably determine:
- Which processes are instrumented using Datadog Tracer's.
- What tracer version and languages are used.
- What services and environments are being instrumented.

### Changes:
  - Add `build.rs` to ease building `ddtrace-native` on MacOS.
  - Store the tracer metadata into a memfd.

Resolves [APMAPI-1067]