chore(llmobs): remove beta auto span linking flag, decorator linking (#13077)
This PR removes the `_DD_LLMOBS_AUTO_SPAN_LINKING_ENABLED` configuration
flag. This is a beta feature that was previously private, but no longer
disabled by default. However removing this flag is not a breaking
change, as this will simply add functionality (add span linking in our
integrations and decorators).

This PR also removes the auto span linking feature for decorator spans.
This was an experimental feature that we started when we first
introduced span linking, but we've decided to remove this for now and
keep span linking limited to our integrations.

In hindsight it should be noted that adding this option to the
`config.py` file was probably a mistake, since it was a private feature
flag that was destined to be removed fairly soon. We should have just
read it directly as an environment variable, but wanted to reiterate
that removing this from the `config.py` file is not a breaking change
(although it would be normally) since this is a private feature flag for
experimental users.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)