perf(profiling): stack-v2 adaptive sampling
We implement a first iteration on adaptive sampling for stack-v2. The
current approach is based on the comparison between the CPU time used by
the sampling thread vs the CPU time used by the rest of the process. The
latter includes the work of our periodic worker processes. However,
these run with period of O(1s) and are mostly on I/O, so we make the
assumption that their contribution towards CPU overhead can be
neglected. The algorithm that we implement aims at reaching an overhead
that is close to the given default threshold by scaling the sampling
interval according to the measured CPU times.