ci: correctly identify base branch for microbenchmarks (#13195)
This PR addresses two separate issues with the microbenchmarks:

1. The `check-big-regressions` job for the microbenchmarks has been
failing on some release branches due to the following silent error in
the `baseline:detect` job:
```
none of the git remotes configured for this repository point to a known GitHub host. To tell gh about a new GitHub host, please use `gh auth login`
```
See example
[here](https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-py/-/jobs/895493368).
As a result, the subsequent `git describe --tags` command would always
be comparing the release branch to the last tag on `main`, not on the
release branch. Adding the `git config`s and `git remote set-url origin`
in the `microbenchmarks.yml` file remediates this.

2. Given that we're moving forward with NO default backporting (except
for 2.21), there will be more cases where a release is tagged on the
`main` branch, not on its own minor release branch like `3.5`. This is
updated in the `detect-baseline.sh` file to reflect that additional
condition. Now if there is no minor release branch, it defaults to main
(see example
[here](https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-py/-/jobs/896300765)).

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)