refactor(profiling): use raw malloc/free, add GIL assertion
For the memalloc traceback scratch buffer pool, use plain malloc and
free rather than PyMem_RawMalloc/PyMem_RawFree. Those functions happen
to just call plain malloc/free, but it's not really guaranteed. Also, it's
important that the pool management functions don't release the GIL,
which is easier to know doesn't happen if we just don't call any Python
C API functions. We may also want to profile the "raw" allocators,
though that's a project for another time.

Related to the above GIL concerns, add an assertion that we hold the GIL
in the pool management functions. We rely on the GIL to protect the pool
state right now. Having the assertions both signals that in the code,
and helps us debug in case we start seeing crashes in the memalloc code.