ci: add manual break glass options for tests and benchmarks (#13222)
This PR adds two variables which can be used to re-run a release process
to allow failures from tests and benchmarks.

This mechanism is meant only as a "break glass" procedure if there is
something in the CI pipeline preventing us from completing a release and
we are willing to accept the risks of not having a green CI run.

Examples could be:

- The release introduces a known/expected performance overhead that we
want to allow
- There is a new flaky test that showed up during this release

This mechanism is very broad, meaning it allows entire child pipelines
to fail and is not targeted to specific jobs. So you cannot say
"benchmark X or test Y are allowed to fail".

In an ideal world these don't need to exist.

- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

- [ ] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

(cherry picked from commit c57516881a846b9954bd19a4394199c1efe6c6e0)